JAMES-3092 Move JMAPUrls to jmap module
