+++ server/protocols/jmap/src/main/java/org/apache/james/jmap/JMAPUrls.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.jmap.http;
+package org.apache.james.jmap;
 
 public interface JMAPUrls {
     String JMAP = "/jmap";
